backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
firebase-admin==6.6.0
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.20
httpx==0.28.1
google-generativeai==0.8.3
cloudinary==1.41.0
Pillow==11.0.0
requests==2.32.3
langchain==1.2.10
langgraph==1.0.9
langchain-core==1.2.16
langchain-google-genai==4.2.1
